Remove unused Scaffold
